requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
rich~=13.9.4
gpt4all~=2.8.2
openai~=1.59.7
transformers~=4.48.0
